FIx for SQL job migration errors
When a job contains the source SQL Server name embedded in a job step the entire job fails to migrate.  Embedding the server name in the job step is not best-practice, but seems to be more common on older versions of SQL Server, especially SQL Server 2005.Considered adding parameter to call the server name swap logic, and possibly letting a different string than the target SQL Server name be the replacement, but wanted to keep the change simple since this is my first.